Revert changes to ga.fish and handle no changes scenario
The global variable `arr` was used by multiple functions, and the new
format caused errors. I reverted the code so global `arr` behaves the same way
prior to my changes and holds the affected files file-names

I also noticed that gs.fish does not handle the case where no changes are
present so I modified the code to handle that that and added an additional test.

Finally, the tests were not running the code present in the library, but
refering to a temp function so I updated that as well.